Add standalone next.js build support

+++ runtimes/node/versions/latest/helpers/next-js/server.sh
@@ -7,13 +7,29 @@ cd /usr/local/server/src/function
 
 source /usr/local/server/helpers/next-js/env.sh
 
-if [ -z "$OPEN_RUNTIMES_START_COMMAND" ]; then
-	# Middleware-style
-	cp ../server-next-js.mjs ./server.mjs
-	START_COMMAND="node ./server.mjs"
+WEBPACK_ENTRYPOINT="./server/webpack-runtime.js"
+TURBOPACK_ENTRYPOINT="./turbopack"
+STANDALONE_ENTRYPOINT="./standalone/server.js"
 
-	# Standalone-style
-	# START_COMMAND="./node_modules/.bin/next start"
+if [ -z "$OPEN_RUNTIMES_START_COMMAND" ]; then
+    # Detect SSR in custom output directory
+	if [ -e "$STANDALONE_ENTRYPOINT" ]; then
+		# Standalone
+		START_COMMAND="node $STANDALONE_ENTRYPOINT"
+	fi
+    
+	ENTRYPOINT="./server/server.js"
+	if [ -e "$WEBPACK_ENTRYPOINT" ] || [ -e "$TURBOPACK_ENTRYPOINT" ]; then
+		# No special export (middleware)
+		cp ../server-next-js.mjs ./server.mjs
+		START_COMMAND="node ./server.mjs"
+	fi
+    
+	# Realistically never happens - build prevents this
+	if [ -z "$START_COMMAND" ]; then
+		echo 'No server found'
+		exit 1
+	fi
 else
 	START_COMMAND="$OPEN_RUNTIMES_START_COMMAND"
 fi
